+++ models/model_nf.py
@@ -70,7 +70,7 @@ def train_model(
 ) -> dict:
 
     # initialize loop
-    model = model.to(config.device)
+    model = model.float().to(config.device)
     model.train()
     size = len(config.train_loader.dataset)
     sum_loss = 0
@@ -154,7 +154,7 @@ def test_model(
 ) -> dict:
 
     # initialize loop
-    model = model.to(config.device)
+    model = model.float().to(config.device)
     model.eval()
     size = len(config.test_loader.dataset)
     sum_loss = 0
