+++ models/model_resnet.py
@@ -78,7 +78,7 @@ def train_model(
 ) -> dict:
 
     # initialize loop
-    model = model.to(config.device)
+    model = model.float().to(config.device)
     model.train()
     size = len(config.train_loader.dataset)
     sum_loss = 0
@@ -163,7 +163,7 @@ def test_model(
 ) -> dict:
 
     # initialize loop
-    model = model.to(config.device)
+    model = model.float().to(config.device)
     model.eval()
     size = len(config.test_loader.dataset)
     sum_loss = 0
